requirements.txt
pytest
selenium
webdriver_manager
webdriver-manager
allure-pytest